Added the main code and a README
The bookmarklet code has been added, along with the README containing the remeinder to "just remember to add the prefix 'javascript:' before you save the code in a bookmark, or run it in the search bar". The README also says that the code is free to use, and fully "open-source", so you can use it at free will.